🤖 Agents/Message Formatter/bot/requirements.txt
python-telegram-bot==21.6
anthropic>=0.25.0
python-dotenv>=1.0.0
